minor code improvements and refactoring #158
