+++ src/main/java/com/github/cameltooling/lsp/internal/completion/CompletionResolverUtils.java
@@ -16,17 +16,30 @@
  */
 package com.github.cameltooling.lsp.internal.completion;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+
+import org.apache.camel.parser.helper.CamelXmlHelper;
 import org.eclipse.lsp4j.CompletionItem;
 import org.eclipse.lsp4j.Position;
 import org.eclipse.lsp4j.Range;
+import org.eclipse.lsp4j.TextDocumentItem;
 import org.eclipse.lsp4j.TextEdit;
+import org.w3c.dom.Node;
 
+import com.github.cameltooling.lsp.internal.instancemodel.CamelURIInstance;
 import com.github.cameltooling.lsp.internal.instancemodel.CamelUriElementInstance;
+import com.github.cameltooling.lsp.internal.instancemodel.PathParamURIInstance;
+import com.github.cameltooling.lsp.internal.parser.ParserXMLFileHelper;
 
 /**
  * @author lheinema
  */
 public class CompletionResolverUtils {
+	
+	private static final List<String> POSSIBLE_DIRECT_REFERENCE = Arrays.asList("direct","direct-vm");
 	
 	private CompletionResolverUtils() {
 		// util class
@@ -48,4 +61,36 @@ public class CompletionResolverUtils {
 			}
 		}
 	}
+	
+	public static boolean isDirectComponentKind(CamelUriElementInstance camelURIInstanceToSearchReference) {
+		return POSSIBLE_DIRECT_REFERENCE.contains(camelURIInstanceToSearchReference.getComponentName());
+	}
+
+	public static String getDirectId(CamelURIInstance camelDirectURIInstance) {
+		Set<PathParamURIInstance> pathParams = camelDirectURIInstance.getComponentAndPathUriElementInstance().getPathParams();
+		if (!pathParams.isEmpty()) {
+			return pathParams.iterator().next().getValue();
+		}
+		return null;
+	}
+
+	public static List<String> retrieveEndpointIDsOfScheme(String scheme, ParserXMLFileHelper xmlFileHelper, TextDocumentItem docItem) throws Exception {
+		List<Node> allEndpoints = xmlFileHelper.getAllEndpoints(docItem);
+		List<String> endpointIDs = new ArrayList<>();
+		for (Node endpoint : allEndpoints) {
+			String uriToParse = CamelXmlHelper.getSafeAttribute(endpoint, "uri");
+			if (uriToParse != null) {
+				CamelURIInstance uriInstance = new CamelURIInstance(uriToParse, endpoint, docItem);
+				if (isDirectComponentKind(uriInstance) && uriInstance.getComponentName().equalsIgnoreCase(scheme)) {
+					String dId = getDirectId(uriInstance);
+					String directValue = String.format("%s:%s", scheme, dId);
+					if (dId != null && dId.trim().length()>0 && !endpointIDs.contains(directValue)) {
+						endpointIDs.add(directValue);
+					}
+				}
+			}
+		}
+		return endpointIDs;
+	}
+
 }
