+++ src/main/java/com/github/cameltooling/lsp/internal/references/ReferencesProcessor.java
@@ -17,13 +17,11 @@
 package com.github.cameltooling.lsp.internal.references;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 
 import org.apache.camel.parser.helper.CamelXmlHelper;
@@ -34,14 +32,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Node;
 
+import com.github.cameltooling.lsp.internal.completion.CompletionResolverUtils;
 import com.github.cameltooling.lsp.internal.instancemodel.CamelURIInstance;
-import com.github.cameltooling.lsp.internal.instancemodel.CamelUriElementInstance;
-import com.github.cameltooling.lsp.internal.instancemodel.PathParamURIInstance;
 import com.github.cameltooling.lsp.internal.parser.ParserXMLFileHelper;
 
 public class ReferencesProcessor {
 	
-	private static final List<String> POSSIBLE_DIRECT_REFERENCE = Arrays.asList("direct","direct-vm");
 	private static final Logger LOGGER = LoggerFactory.getLogger(ReferencesProcessor.class);
 	private ParserXMLFileHelper parserXMLFileHelper = new ParserXMLFileHelper();
 	private TextDocumentItem textDocumentItem;
@@ -56,7 +52,7 @@ public class ReferencesProcessor {
 			try {
 				String camelComponentUri = parserXMLFileHelper.getCamelComponentUri(textDocumentItem, position);
 				CamelURIInstance camelURIInstanceToSearchReference = parserXMLFileHelper.createCamelURIInstance(textDocumentItem, position, camelComponentUri);
-				if (isDirectComponentKind(camelURIInstanceToSearchReference)) {
+				if (CompletionResolverUtils.isDirectComponentKind(camelURIInstanceToSearchReference)) {
 					Map<CamelURIInstance, Node> allCamelUriInstances = retrieveAllEndpoints();
 					return CompletableFuture.completedFuture(findReferences(camelURIInstanceToSearchReference, allCamelUriInstances));
 				}
@@ -69,7 +65,7 @@ public class ReferencesProcessor {
 
 	private List<Location> findReferences(CamelURIInstance camelURIInstanceToSearchReference, Map<CamelURIInstance, Node> allCamelUriInstance) {
 		List<Location> references = new ArrayList<>();
-		String directId = getDirectId(camelURIInstanceToSearchReference);
+		String directId = CompletionResolverUtils.getDirectId(camelURIInstanceToSearchReference);
 		if (directId != null && !directId.isEmpty()) {
 			for (Entry<CamelURIInstance, Node> entry : allCamelUriInstance.entrySet()) {
 				CamelURIInstance camelURIInstance = entry.getKey();
@@ -82,10 +78,10 @@ public class ReferencesProcessor {
 	}
 
 	private boolean isReference(CamelURIInstance camelURIInstanceToSearchReference, String directId, CamelURIInstance camelURIInstance) {
-		return isDirectComponentKind(camelURIInstance)
+		return CompletionResolverUtils.isDirectComponentKind(camelURIInstance)
 				&& (camelURIInstanceToSearchReference.isProducer() && !camelURIInstance.isProducer()
 						|| !camelURIInstanceToSearchReference.isProducer() && camelURIInstance.isProducer())
-				&& directId.equals(getDirectId(camelURIInstance));
+				&& directId.equals(CompletionResolverUtils.getDirectId(camelURIInstance));
 	}
 
 	private Map<CamelURIInstance, Node> retrieveAllEndpoints() throws Exception {
@@ -98,36 +94,5 @@ public class ReferencesProcessor {
 			}
 		}
 		return allCamelUriInstance;
-	}
-	
-	private static String getDirectId(CamelURIInstance camelDirectURIInstance) {
-		Set<PathParamURIInstance> pathParams = camelDirectURIInstance.getComponentAndPathUriElementInstance().getPathParams();
-		if (!pathParams.isEmpty()) {
-			return pathParams.iterator().next().getValue();
-		}
-		return null;
-	}
-
-	public static boolean isDirectComponentKind(CamelUriElementInstance camelURIInstanceToSearchReference) {
-		return POSSIBLE_DIRECT_REFERENCE.contains(camelURIInstanceToSearchReference.getComponentName());
-	}
-
-	public static List<String> retrieveEndpointIDsOfScheme(String scheme, ParserXMLFileHelper xmlFileHelper, TextDocumentItem docItem) throws Exception {
-		List<Node> allEndpoints = xmlFileHelper.getAllEndpoints(docItem);
-		List<String> endpointIDs = new ArrayList<>();
-		for (Node endpoint : allEndpoints) {
-			String uriToParse = CamelXmlHelper.getSafeAttribute(endpoint, "uri");
-			if (uriToParse != null) {
-				CamelURIInstance uriInstance = new CamelURIInstance(uriToParse, endpoint, docItem);
-				if (isDirectComponentKind(uriInstance) && uriInstance.getComponentName().equalsIgnoreCase(scheme)) {
-					String dId = getDirectId(uriInstance);
-					String directValue = String.format("%s:%s", scheme, dId);
-					if (dId != null && dId.trim().length()>0 && !endpointIDs.contains(directValue)) {
-						endpointIDs.add(directValue);
-					}
-				}
-			}
-		}
-		return endpointIDs;
 	}
 }
